+++ core-bundle/src/Resources/contao/dca/tl_content.php
@@ -533,7 +533,7 @@ $GLOBALS['TL_DCA']['tl_content'] = array
 		),
 		'orderSRC' => array
 		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_content']['orderSRC'],
+			'label'                   => &$GLOBALS['TL_LANG']['MSC']['sortOrder'],
 			'sql'                     => "blob NULL"
 		),
 		'useHomeDir' => array