+++ core-bundle/src/Resources/contao/library/Contao/Controller.php
@@ -1788,6 +1788,38 @@ abstract class Controller extends \System
 			}
 		}
 
+		// Order the enclosures
+		if (!empty($arrItem['orderEnclosure']))
+		{
+			$tmp = \StringUtil::deserialize($arrItem['orderEnclosure']);
+
+			if (!empty($tmp) && \is_array($tmp))
+			{
+				// Remove all values
+				$arrOrder = array_map(function () {}, array_flip($tmp));
+
+				// Move the matching elements to their position in $arrOrder
+				foreach ($arrEnclosures as $k=>$v)
+				{
+					if (array_key_exists($v['uuid'], $arrOrder))
+					{
+						$arrOrder[$v['uuid']] = $v;
+						unset($arrEnclosures[$k]);
+					}
+				}
+
+				// Append the left-over enclosures at the end
+				if (!empty($arrEnclosures))
+				{
+					$arrOrder = array_merge($arrOrder, array_values($arrEnclosures));
+				}
+
+				// Remove empty (unreplaced) entries
+				$arrEnclosures = array_values(array_filter($arrOrder));
+				unset($arrOrder);
+			}
+		}
+
 		$objTemplate->enclosure = $arrEnclosures;
 	}